pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dspd"
version = "0.1.0"
description = "DevOps Solution Product Development — Plane integration for the OpenClaw Fleet"
readme = "README.md"
license = { text = "AGPLv3" }
requires-python = ">=3.11"
authors = [
    { name = "OpenClaw Fleet", email = "fleet@openclaw.ai" },
]
keywords = ["plane", "project-management", "fleet", "openclaw"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: GNU Affero General Public License v3",
]

dependencies = [
    # HTTP client (async)
    "httpx>=0.27.0",
    # Data validation and models
    "pydantic>=2.7.0",
    # CLI framework
    "typer>=0.12.0",
    "rich>=13.7.0",
    # Environment / config
    "python-dotenv>=1.0.0",
    # YAML config parsing
    "pyyaml>=6.0",
    # Structured logging
    "structlog>=24.1.0",
    # Webhook receiver (ASGI)
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",    # Mock httpx in tests
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-requests",
]

[project.scripts]
dspd = "dspd.cli.main:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["dspd"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: fast unit tests (no network)",
    "integration: requires running Plane instance",
    "golden: golden fixture tests",
]

[tool.hatch.build.targets.wheel]
packages = ["dspd", "fleet"]
